The project is almost ready
The (input @visualizable request) line from AmILabServer.adf2 may cause
problems.
The active input mechanism implemented in the AmILabGui may not work.
If these two problems are fixed, sclaim_app is ready for deployment on
three machines (one of them is both server and client).
The daqs used are 1 (near the exit, but not on the same wall), 2
(windows, under the air conditioning) and 5 (wall opposed to 1, under
the air conditioning).